LPD-53896 Rename directory, no functional changes
