+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/dnd/DragLayer.tsx
