+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/views/list/List.tsx
@@ -14,8 +14,8 @@ import React, {forwardRef, useContext} from 'react';
 import FrontendDataSetContext from '../../FrontendDataSetContext';
 import Actions from '../../actions/Actions';
 import ImageRenderer from '../../cell_renderers/ImageRenderer';
-import FDSDndProvider from '../../drag_drop/FDSDndProvider';
-import useFDSDrop from '../../drag_drop/useFDSDrop';
+import FDSDndProvider from '../../dnd/FDSDndProvider';
+import useFDSDrop from '../../dnd/useFDSDrop';
 import {getLocalizedValue} from '../../utils/getLocalizedValue';
 import {IHeader, IListSchema, IListTitleRenderer} from '../../utils/types';
